backend/requirements.txt
# Core server
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic>=2.12.0

# Database
motor==3.3.1
pymongo==4.5.0

# Image & QR/Barcode
pillow>=11.3.0
qrcode>=8.2
python-barcode>=0.16.1

# AI integrations (optional - install if needed)
openai>=1.99.9
google-generativeai>=0.8.5
# emergentintegrations==0.1.0  # Optional - not publicly available

# CORS
starlette==0.37.2
